[Feature][transform] transform support explode
